requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
Werkzeug==2.2.3
pdfplumber==0.10.3
spacy==3.7.2
transformers==4.36.2
torch==2.2.0
nltk==3.8.1
PyPDF2==3.0.1
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2
